Added support for building for the Intel compiler using CHPL_HOST_PLATFORM linux-intel.  When David comes to town, he and I should talk about potentially adding some sort of CHPL_HOST_COMPILER variable that can be used to select a particular compiler suite, given that he has similar issues on XT (or maybe he doesn't).
Made our code more compatible with the Intel compiler when using the
-Wall and -Werror flags (though I suppress a number of warnings that I
don't think we want to live by, noted below).  From most to least
important/interesting, the changes this resulted in are:

- moved all pass names into the file passes.h (out of passlist.h
  and/or the file itself) so that there was a single place that pass
  names could be created (passlist.h wasn't a very good place because
  it's more difficult to #include).

- moved runtime prototypes of routines in the generated code into a
  common header file -- chplcgfns.h.  Previously, these were defined
  in the C files (generally a bad organizational idea), and scattered
  in other header files.

- changed David's destructure_arg_tuples.chpl test to not return an
  integer from main, as this wasn't what the test was written to test,
  and because it causes problems and raises semantic questions (this
  wasn't caught earlier because the extern of chpl_main was never
  compared against what we generated).  Added a new semantic future
  test to capture this behavior without breaking anything.  (Would've
  run this past David, but he's on a plane).

- as part of this, fixed the prototypes of the cid2*() routines so
  that they correctly refer to a _class_id rather than an int64.  To
  do this, I reorganized the code generation of _chpl_header slightly
  so that the enum is created first, then it #defines a macro
  indicating that we're in the generated code, then it #includes
  stdchpl.h (previously, stdchpl.h was included first, then the enum
  was defined).  In chplcgfns.h, if we're not in the generated code,
  we now define a bogus version of _class_id so that the enum can be
  referred to in the runtime (even though it's values cannot be
  accessible -- nor should they be).

- fixed up various function prototyping problems by:
  - replacing extern declarations in source files with #inclusion of
    appropriate headers
  - prototyping external functions in header files if they weren't
  - making other functions static
  - in one case making a function (tag_generic()) into a method, as it
    was very method-like and there wasn't a good place to drop a
    prototype
  - in another case, made paramMap into a function argument rather
    than referring to a global variable non-locally

- removed unused function arguments (but turned off warning for future
  purposes because there are a bunch of false matches in the case of
  functions that must have a certain signature -- e.g., method
  overloading or passing function pointers).

- fixed up printf and scanf formats
  - perhaps most interestingly, changed our int8 typedefs from "char"
    to "signed char" because C is vague about whether char is signed
    or unsigned

- cleaned up lexer/parser stuff a bit to avoid warnings/errors

- removed dead static functions

- removed an extraneous semicolon

- removed dead variables

- removed unnecessary local prototypes


The warnings that I'm currently suppressing are:

304  : requires inherited classes to specify access control (e.g. public)
593  : variable was set but never used
810  : numeric conversion may lose significant bits (e.g., long double->float)
869  : warns about arguments that are never used
981  : operands are evaluated in unspecified order (mysterious vec.h problems)
1572 : floating point equality and inequality comparisons are unreliable
1599 : one variable shadows another


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@12111 3a8e244f-b0f2-452b-bcba-4c88e055c3ca

+++ compiler/ifa/num.cpp
@@ -109,13 +109,13 @@ sprint_imm(char *str, Immediate &imm) {
     case NUM_KIND_UINT: {
       switch (imm.num_index) {
         case INT_SIZE_1: 
-          res = sprintf(str, "%u", imm.v_bool); break;
+          res = sprintf(str, "%u", (unsigned)imm.v_bool); break;
         case INT_SIZE_8: 
-          res = sprintf(str, "%u", imm.v_uint8); break;
+          res = sprintf(str, "%u", (unsigned)imm.v_uint8); break;
         case INT_SIZE_16:
-          res = sprintf(str, "%u", imm.v_uint16); break;
+          res = sprintf(str, "%u", (unsigned)imm.v_uint16); break;
         case INT_SIZE_32:
-          res = sprintf(str, "%u", imm.v_uint32); break;
+          res = sprintf(str, "%u", (unsigned)imm.v_uint32); break;
         case INT_SIZE_64:
           res = sprintf(str, "%llu", imm.v_uint64); break;
         default: INT_FATAL("Unhandled case in switch statement");
@@ -179,13 +179,13 @@ fprint_imm(FILE *fp, Immediate &imm) {
     case NUM_KIND_UINT: {
       switch (imm.num_index) {
         case INT_SIZE_1: 
-          res = fprintf(fp, "%u", imm.v_bool); break;
+          res = fprintf(fp, "%u", (unsigned)imm.v_bool); break;
         case INT_SIZE_8: 
-          res = fprintf(fp, "%u", imm.v_uint8); break;
+          res = fprintf(fp, "%u", (unsigned)imm.v_uint8); break;
         case INT_SIZE_16:
-          res = fprintf(fp, "%u", imm.v_uint16); break;
+          res = fprintf(fp, "%u", (unsigned)imm.v_uint16); break;
         case INT_SIZE_32:
-          res = fprintf(fp, "%u", imm.v_uint32); break;
+          res = fprintf(fp, "%u", (unsigned)imm.v_uint32); break;
         case INT_SIZE_64:
           res = fprintf(fp, "%llu", imm.v_uint64); break;
         default: INT_FATAL("Unhandled case in switch statement");
